refactor(template-compiler): remove BindingCommand.prototype.compile
chore(tests): fix i18n tests
